requirements.txt
os
json
requests
logging
sqlalchemy
datetime
python-dotenv
psycopg2
flask_sqlalchemy
flask